Refactor drawing of shapes

+++ src/resources.rs
@@ -29,6 +29,6 @@ pub struct KillProgram(pub bool);
 pub struct GameMap(pub HashMap<u32, f64>);
 
 pub struct GameMap2(
-    pub  [[i8; (settings::NUMBER_OF_CELLS_HIGH as usize)];
+    pub  [[i16; (settings::NUMBER_OF_CELLS_HIGH as usize)];
         (settings::NUMBER_OF_CELLS_WIDE as usize)],
 );
